+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The Apache Spark-based engine now uses INT64 to represent timestamps when
   reading and writing Apache Parquet files for compatibility with the H2-based
   engine.
+- Both engines now use the same format when serializing timestamps to JSON.
 
 ### Deprecated
 ### Removed